+++ BUILD.gn
@@ -72,7 +72,7 @@ v8_source_set("deno_nosnapshot") {
     "src/deno_internal.h",
     "src/file_util.cc",
     "src/file_util.h",
-    "src/include/deno.h",
+    "src/deno.h",
   ]
   deps = [
     "third_party/v8:v8_monolith",
