+++ src/binding.cc
@@ -29,7 +29,7 @@ IN THE SOFTWARE.
 #include "third_party/v8/src/base/logging.h"
 
 #include "./deno_internal.h"
-#include "include/deno.h"
+#include "deno.h"
 
 namespace deno {
