+++ src/snapshot_creator.cc
@@ -3,7 +3,7 @@
 // Hint: --trace_serializer is a useful debugging flag.
 #include "deno_internal.h"
 #include "file_util.h"
-#include "include/deno.h"
+#include "deno.h"
 #include "third_party/v8/include/v8.h"
 #include "third_party/v8/src/base/logging.h"
